確認 CP 審核邏輯

+++ server/app/models/die_data.py
@@ -1,7 +1,15 @@
 from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey, Numeric, Index
 from sqlalchemy.orm import relationship
 
+from sqlalchemy.dialects.mysql import DOUBLE
+
 from app.database import Base
+
+# DOUBLE, not NUMERIC: CP data spans volts down to picoamps, and a
+# fixed-point scale wide enough for both is impractical. NUMERIC(15,6)
+# rounded every nanoamp reading and limit to exactly zero. See
+# migration 016.
+
 
 
 class DieData(Base):
@@ -26,7 +34,7 @@ class ElectricalValue(Base):
     id = Column(BigInteger, primary_key=True, autoincrement=True)
     die_id = Column(BigInteger, ForeignKey("die_data.id"), nullable=False)
     param_name = Column(String(50), nullable=False)
-    value = Column(Numeric(15, 6))
+    value = Column(DOUBLE)
 
     die = relationship("DieData", back_populates="electrical_values")
